+++ src/signs/ui/search/SignsSearchForm.tsx
@@ -1,71 +1,22 @@
 import React, { useState } from 'react'
 import { stringify } from 'query-string'
-import { Button, Col, Form, FormControl, Popover, Row } from 'react-bootstrap'
+import { Button, Col, Form, FormControl, Row } from 'react-bootstrap'
 import { RouteComponentProps, withRouter } from 'react-router-dom'
-import _ from 'lodash'
 import HelpTrigger from 'common/HelpTrigger'
 import { SignQuery } from 'signs/domain/Sign'
 import replaceTransliteration from 'fragmentarium/domain/replaceTransliteration'
+import SignsSearchHelpPopover from 'signs/ui/search/SignsSearchHelpPopover'
 
 interface Props extends RouteComponentProps {
   signQuery: SignQuery
   sign: string | undefined
 }
-
-function SignsSearchHelp(): JSX.Element {
-  const Section = ({ label, text }: { label: string; text: string }) => (
-    <li>
-      <Row>
-        <Col className="pr-1 mr-0 Help__list-name">{label}</Col>
-        <Col xs="auto" className="pl-0 ml-0">
-          =
-        </Col>
-        <Col className="pl-0">{text}</Col>
-      </Row>
-    </li>
-  )
-  const signsSearchHelpList = [
-    ['MZL', 'R. Borger, *Mesopotamisches Zeichenlexikon* (Münster, ²2010).'],
-    [
-      'ŠL/MÉA',
-      'A. Deimel, *Šumerisches Lexikon* (Rom, 1925/1950) / R. Labat, *Manuel d’épigraphie akkadienne* (Paris, ⁶1988).',
-    ],
-    [
-      'ABZ',
-      'R. Borger, *Assyrisch-babylonische Zeichenliste* (Neukirchen-Vluyn, ⁴1988).',
-    ],
-    [
-      'OBZL',
-      'C. Mittermayer, *Altbabylonische Zeichenliste der sumerisch-literarischen Texte* (Göttingen, 2006).',
-    ],
-    [
-      'KWU',
-      'N. Schneider, *Die Keilschriftzeichen der Wirtschaftsurkunden von Ur III* (Rom, 1935).',
-    ],
-    [
-      'LAK',
-      'A. Deimel, *Liste der archaischen Keilschriftzeichen* (Leipzig, 1922).',
-    ],
-    [
-      'HZL',
-      'Ch. Rüster; E. Neu, *Hethitisches Zeichenlexikon* (Wiesbaden, 1989).',
-    ],
-  ]
-  return (
-    <Popover
-      id={_.uniqueId('SignsSearchHelp-')}
-      title="Search transliterations"
-      className="Help__popover"
-    >
-      <Popover.Content>
-        <ul>
-          {signsSearchHelpList.map((help, index) => (
-            <Section key={index} label={help[0]} text={help[1]} />
-          ))}
-        </ul>
-      </Popover.Content>
-    </Popover>
-  )
+function parseValue(sign: string): { value: string; subIndex: number } {
+  const match = sign.match(/^([^\d]+)(\d*)$/)
+  return {
+    value: match ? replaceTransliteration(match[1]) : '',
+    subIndex: match && match[2] ? parseInt(match[2]) : 1,
+  }
 }
 
 function SignsSearchForm({ sign, signQuery, history }: Props): JSX.Element {
@@ -74,14 +25,6 @@ function SignsSearchForm({ sign, signQuery, history }: Props): JSX.Element {
   const [unnormalizedSignQuery, setUnnormalizedSignQuery] = useState(
     `${signQuery.value}${signQuery.subIndex ? signQuery.subIndex : ''}`
   )
-
-  const parseValue = (sign: string): { value: string; subIndex: number } => {
-    const match = sign.match(/^([^\d]+)(\d*)$/)
-    return {
-      value: match ? replaceTransliteration(match[1]) : '',
-      subIndex: match && match[2] ? parseInt(match[2]) : 1,
-    }
-  }
 
   const query = (event) => {
     event.preventDefault()
@@ -162,7 +105,7 @@ function SignsSearchForm({ sign, signQuery, history }: Props): JSX.Element {
       <Form.Group>
         <Row>
           <Col xs={2} className="ml-3">
-            <HelpTrigger overlay={SignsSearchHelp()} />
+            <HelpTrigger overlay={SignsSearchHelpPopover()} />
           </Col>
           <Col className="pl-0 ml-0 mr-0 pr-0">
             <Form.Control
